InstrumentationScope to include attributes in hash and eq check (#2701)
